+++ R/loading.R
@@ -572,7 +572,7 @@ save_file_multi <- function(data_frame_list,
 #'
 #' # Load file
 #' my_fst <- load_file(path = dirname(fst_file),
-#'                     file = basename(rds_file))
+#'                     file = basename(fst_file))
 #' my_rds <- load_file(path = dirname(rds_file),
 #'                     file = basename(rds_file))
 #'
@@ -580,14 +580,22 @@ save_file_multi <- function(data_frame_list,
 #' # NOTE: Variable names can be written case insensitive. Meaning if a variable
 #' #       is stored as "age" and you write "AGE" in keep, the function will find
 #' #       the variable and rename it to "AGE".
-#' my_fst_keep<- load_file(path = dirname(fst_file),
-#'                         file = basename(rds_file),
-#'                         keep = c(AGE, INCOME_class, State, weight))
+#' my_fst_keep <- load_file(path = dirname(fst_file),
+#'                          file = basename(fst_file),
+#'                          keep = c(AGE, INCOME_class, State, weight))
 #'
 #' # Load file and subset observations
 #' my_fst_where <- load_file(path  = dirname(fst_file),
-#'                           file  = basename(rds_file),
+#'                           file  = basename(fst_file),
 #'                           where = sex == 1 & age > 65)
+#'
+#' # Load file and directly rename variables
+#' my_fst_rename <- load_file(path = dirname(fst_file),
+#'                            file = basename(fst_file),
+#'                            keep = list(AGE          = SingleAges,
+#'                                        INCOME_class = NetIncome,
+#'                                        State        = GermanState,
+#'                                        weight       = ExtraFactor))
 #'
 #' @rdname save_load
 #'