+++ NEWS.md
@@ -19,6 +19,7 @@
 * `any_table()`: When using `by` variables the special keyword `[by_var]` is now replaced with the actual `by` value in the titles and footnotes. (01.08.2026)
 * `frequencies()`, `crosstabs()`: Now also export the Excel workbook. (02.08.2026)
 * `frequencies()`, `crosstabs()`: When using `by` variables the special keyword `[by_var]` is now replaced with the actual `by` value in the titles and footnotes. (01.08.2026)
+* `where.()`: Added a html rendered view of the data frame as fallback, if the new window through `utils::View()` can't be opened. (03.08.2026)
 
 ### Changed functionality
 
@@ -51,6 +52,8 @@
 * Cleanup of a partial matches. (22.07.2026, thanks to @JanMarvin and @olivroy)
 * `stack_data()`: Swapped out a copypasta error message with the right one. (22.07.2026)
 * `combine_into_workbook()`: Added example on how to save a file. (03.08.2026)
+* `dummy_data()`: `NUTS2` variable is now stored as an integer. (03.08.2026)
+* `load_file()`: Added example which shows how to directly rename variables. (03.08.2026)
 
 # qol 1.3.3